% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package.R
\docType{package}
\name{PTAdcbcomparison3-package}
\alias{PTAdcbcomparison3}
\alias{PTAdcbcomparison3-package}
\title{PTAdcbcomparison3: A Package Skeleton for Comparative Effectiveness Studies}
\description{
A skeleton package, to be used as a starting point when implementing comparative effect studies.
}
\keyword{internal}
