% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package.R
\docType{package}
\name{PTAdcbcomparison2-package}
\alias{PTAdcbcomparison2}
\alias{PTAdcbcomparison2-package}
\title{PTAdcbcomparison2: A Package Skeleton for Comparative Effectiveness Studies}
\description{
A skeleton package, to be used as a starting point when implementing comparative effect studies.
}
\keyword{internal}
